Rename mod kas_view::data_traits -> data_clerk
